+++ DHARMa/R/compatibility.R
@@ -229,6 +229,41 @@ getFamily <- function (object, ...) {
   UseMethod("getFamily", object)
 }
 
+#' Get model data
+#'
+#' Wrapper to get the data that was used to fit a model.
+#'
+#' @param object a fitted model.
+#' @param ... additional parameters to be passed on.
+#'
+#' @seealso [getObservedResponse], [getSimulations], [getRefit], [getFixedEffects], [getFitted]
+#'
+#' @author Florian Hartig
+#' @export
+getData <- function (object, ...) {
+  UseMethod("getData", object)
+}
+
+
+
+#' Get predictor names
+#'
+#' Wrapper to get the names of the predictors used in the model
+#'
+#' @param object a fitted model.
+#' @param ... additional parameters to be passed on.
+#'
+#' @seealso [getObservedResponse], [getSimulations], [getRefit], [getFixedEffects], [getFitted]
+#'
+#' @author Florian Hartig
+#' @export
+getPredictorNames <- function (object, ...) {
+  UseMethod("getPredictorNames", object)
+}
+
+
+
+
 # nObs
 
 # default -----------------------------------------------------------------
@@ -369,6 +404,17 @@ getFamily.default <- function (object,...){
   family(object)
 }
 
+#' @rdname getData
+#' @export
+getData.default <- function (object,...){
+  eval(object$call$data, envir = environment(formula(object)))
+}
+
+#' @rdname getPredictorNames
+#' @export
+getPredictorNames.default <- function (object,...){
+  colnames(model.frame(object))[-1]
+}
 
 
 ######### LM #############
@@ -533,6 +579,13 @@ getSimulations.merMod <- function (object, nsim = 1, simulateREs = c("conditiona
   }
 
   return(out)
+}
+
+
+#' @rdname getData
+#' @export
+getData.merMod <- function (object, ...){
+  eval(object@call$data, envir = environment(formula(object)))
 }
 
 
@@ -782,6 +835,14 @@ getFitted.MixMod <- function (object,...){
 getResiduals.MixMod <- function (object,...){
   residuals(object, type = "subject_specific")
 }
+
+
+#' @rdname getPredictorNames
+#' @export
+getPredictorNames.MixMod <- function (object,...){
+  c(colnames(model.frame(object))[-1], colnames(object$id))
+}
+
 
 ####### phylolm / phyloglm #########