+++ DHARMa/R/plots.R
@@ -152,8 +152,8 @@ plotQQunif <- function(simulationOutput, testUniformity = TRUE, testOutliers = T
 #' The function creates a generic residual plot with either spline or quantile regression to highlight patterns in the residuals. Outliers are marked in star/asterisk form and highlighted in red if the outlier test is significant.
 #'
 #' @param simulationOutput an object, usually a DHARMa object, from which residual values can be extracted. Alternatively, a vector with residuals or a fitted model can be provided, which will then be transformed into a DHARMa object.
-#' @param form optional predictor against which the residuals should be plotted. Default is to use the predicted(simulationOutput).
-#' @param quantreg whether to perform a quantile regression based on [testQuantiles] or a smooth spline around the mean. Default NULL chooses T for nObs < 2000, and F otherwise.
+#' @param form optional predictor(s) against which the residuals should be plotted. Default is to use the predicted(simulationOutput). Recommended to specify as formula. Use \code{form = ~.} to plot against all predictors. See details.
+#' @param quantreg whether to perform a quantile regression based on [testQuantiles] or a smooth spline around the mean. Default NULL chooses T for nObs < 10000, and F otherwise.
 #' @param rank if T, the values provided in form will be rank transformed. This will usually make patterns easier to spot visually, especially if the distribution of the predictor is skewed. If form is a factor, this has no effect.
 #' @param asFactor should a numeric predictor provided in form be treated as a factor. Default is to choose this for < 10 unique values, as long as enough predictions are available to draw a boxplot.
 #' @param smoothScatter if T, a smooth scatter plot will plotted instead of a normal scatter plot. This makes sense when the number of residuals is very large. Default NULL chooses T for nObs > 10000, and F otherwise.
@@ -168,9 +168,9 @@ plotQQunif <- function(simulationOutput, testUniformity = TRUE, testOutliers = T
 #'
 #' Overdispersion typically manifests itself as Q1 (0.25) deviating towards 0 and Q3 (0.75) deviating towards 1. Heteroskedasticity manifests itself as non-parallel quantile lines. To diagnose heteroskedasticity and overdispersion, it can be helpful to additionally plot the absolute deviation of the residuals from the mean expectation of 0.5, using the option absoluteDeviation = T. In this case, we would again expect Q1-Q3 quantile lines at 0.25, 0.5, 0.75, but greater dispersion (also locally in the case of heteroskedasticity) always manifests itself in deviations towards 1.
 #'
-#' The quantile regression can take some time to calculate, especially for larger datasets. For this reason, quantreg = F can be set to generate a smooth spline instead. This is the default for n > 2000.
+#' The quantile regression can take some time to calculate, especially for larger datasets. For this reason, quantreg = F can be set to generate a smooth spline instead. This is the default for n > 10000.
 #'
-#' If form is a factor, a boxplot will be plotted instead of a scatter plot. The distribution for each factor level should be uniformly distributed, so the box should go from 0.25 to 0.75, with the median line at 0.5 (within-group). To test if deviations from those expectations are significant, KS-tests per group and a Levene test for homogeneity of variances is performed. See [testCategorical] for details.
+#' If form is specified as a formula, e.g. \code{form = ~ your_predictor}, NAs will be handled automatically (recommended). If \code{form = ~.} a separate plot for each predictor in the model is produced. If \code{form = ~ predictor1 + predictor2}, a separate plot for each predictor is produced. If an additional grouping variable is specified, e.g. \code{form = ~ predictor|group == "group_level"}, a residual plot for the specified grouping level will be created. If form is **not** a formula, e.g. \code{form = data$predictor}, NAs are **not** handled automatically. For phyr and gamm4$mer models, you need to specify form in this way. If the predictor is a factor, a boxplot will be plotted instead of a scatter plot. The distribution for each factor level should be uniformly distributed, so the box should go from 0.25 to 0.75, with the median line at 0.5 (within-group). To test if deviations from those expectations are significant, KS-tests per group and a Levene test for homogeneity of variances is performed. See [testCategorical] for details.
 #'
 #' @note If nObs > 10,000, the scatter plot is replaced by graphics::smoothScatter().
 #'
@@ -193,8 +193,8 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
   yAxis = ifelse(absoluteDeviation == TRUE, "Residual spread [2*abs(res - 0.5)]", "DHARMa residual")
   a$ylab = checkDots("ylab", yAxis , ...)
   a$xlab = checkDots("xlab", ifelse(is.null(form), "Model predictions",
-                                    gsub(".*[$]","",deparse(substitute(form)))), ...)
-  if(rank == TRUE) a$xlab = paste(a$xlab, "(rank transformed)")
+                                    gsub(".*[$]|^~","",deparse(substitute(form)))), ...)
+  a$main = checkDots("main", NULL, ...)
 
   simOut <- simulationOutput
   simulationOutput = ensureDHARMa(simulationOutput, convert = TRUE)
@@ -206,7 +206,93 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
 
   if(inherits(form, "DHARMa"))stop("DHARMa::plotResiduals > argument form cannot be of class DHARMa. Note that the syntax of plotResiduals has changed since DHARMa 0.3.0. See ?plotResiduals.")
 
-  pred = ensurePredictor(simulationOutput, form)
+  ##### form #####
+  checkForm = substitute(form)
+  if(!is.null(checkForm)) {
+    if (!inherits(form, "formula")) {
+      predictor = form # go back to old default
+      group = NULL
+      if(is.null(predictor)) stop("DHARMA: unable to find specified predictor variable.")
+
+    } else {
+
+      if(simulationOutput$modelClass == "communityPGLMM") stop("DHARMa: to use plotResiduals with phyr models, please return to the old DHARMa default, e.g. form = your_data$your_predictor.")
+
+      if(form == "~.") {
+        predictorNames = getPredictorNames(simulationOutput$fittedModel)
+        return(plotResidualsAll(
+          simulationOutput = simulationOutput,
+          predictorNames = predictorNames,
+          form = form,
+          quantreg = quantreg,
+          rank = rank,
+          asFactor = asFactor,
+          smoothScatter = smoothScatter,
+          quantiles = quantiles,
+          absoluteDeviation = absoluteDeviation,
+          ...
+          ))}
+
+      # if there is more than 1 predictor, check if form is specified either
+      # with "|" (plot residuals within a certain group)
+      # or "+" (plot residuals for multiple predictors in separate panels via plotResidualsAll)
+
+      predictors = getFormulaPredictors(simulationOutput = simulationOutput, formula = form)
+
+      if(length(predictors) > 1) {
+        operators = unlist(strsplit(as.character(form[[2]]), split = character(0)))
+
+        if("|" %in% operators && "+" %in% operators) stop("DHARMa: Detected both + and | as operators. Please choose one or the other. See ?plotResiduals.")
+
+        # for formulas with +, create a separate residual plot for each predictor
+        if("+" %in% operators) {
+          predictorNames = names(predictors)
+
+          return(plotResidualsAll(
+            simulationOutput = simulationOutput,
+            predictorNames = predictorNames,
+            form = form,
+            quantreg = quantreg,
+            rank = rank,
+            asFactor = asFactor,
+            smoothScatter = smoothScatter,
+            quantiles = quantiles,
+            absoluteDeviation = absoluteDeviation,
+            ...
+          ))
+
+        }
+
+        # for formulas with |, create a residual plot for the specified grouping level
+        predictor = predictors[[1]]
+        group = predictors[[2]]
+
+        # if form has just one predictor, make the standard plot for that predictor
+      } else {
+        predictor = predictors[[1]]
+        group = NULL
+      }
+
+      if(is.null(predictor)) stop("DHARMA: unable to extract specified predictor variable from the data frame that was used to fit the model. Further note: if you are using a gamm4 model, please use form = your_data$your_predictor instead of form = ~ your_predictor.")
+    }
+  } else {
+    predictor = NULL
+    group = NULL
+  }
+
+  pred = ensurePredictor(simulationOutput, predictor)
+  group = if(!is.null(group)) {ensurePredictor(simulationOutput, group)}
+
+  # if group is not NULL, get predictor and residuals belonging to specified group level
+  if(!is.null(group)) {
+    tryCatch({groupLevel = eval(form[[2]][[3]][[3]])}, error = function(e) stop("DHARMa: Please specify a certain group level to be plotted."))
+    pred = pred[group == groupLevel]
+    if(length(pred) == 0) {stop("DHARMa: Unable to find specified group level.")}
+    res = simulationOutput$scaledResiduals[group == groupLevel]
+  }
+
+
+
 
   ##### Rank transform and factor conversion#####
 
@@ -215,6 +301,7 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
     if (rank == TRUE){
       pred = rankTransform(pred)
       a$xlim = checkDots("xlim", c(0,1), ...)
+      a$xlab = paste(a$xlab, "(rank transformed)")
     }
 
     nuniq = length(unique(pred))
@@ -225,7 +312,13 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
 
   ##### Residual scatter plots #####
 
-  if(is.null(quantreg)) if (length(res) > 2000) quantreg = FALSE else quantreg = TRUE
+  if(is.null(quantreg)) {
+    if (length(res) > 10000) {
+      quantreg = FALSE
+      message("plotResiduals() for datasets larger than 10,000 displays a spline instead of quantile regression lines for the sake of speed. To switch back to quantile regression lines, use the argument quantreg = T and have in mind that the function may take a while to compute and display the lines.")
+      }
+    else quantreg = TRUE
+  }
 
   switchScatter = 10000
   if(is.null(smoothScatter)) if (length(res) > switchScatter) smoothScatter = TRUE else smoothScatter = FALSE
@@ -236,8 +329,15 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
 
   # categorical plot
   if(is.factor(pred)){
-    testCategorical(simulationOutput = simulationOutput, catPred = pred,
-                    quantiles = quantiles)
+    if(is.null(group)) {
+      testCategorical(simulationOutput = simulationOutput, catPred = pred,
+                      quantiles = quantiles, xlab = a$xlab, ...)
+    } else {
+      simulationOutput$scaledResiduals = res
+      testCategorical(simulationOutput = simulationOutput, catPred = pred,
+                      quantiles = quantiles, xlab = a$xlab, ...)
+    }
+
   } else{
 
     # color/shape outliers according to significance of outlier test, related to issue #453
@@ -258,7 +358,7 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
     if (smoothScatter == TRUE) {
       do.call(graphics::smoothScatter, append(list(x = pred, y = res ,
                                                    ylim = c(0,1), axes = FALSE,
-                                                   colramp = colorRampPalette(c("white", "darkgrey"))),a))
+                                                   colramp = colorRampPalette(c("white", "darkgrey"))),a[names(a) != "main"]))
       points(pred[outlier], res[outlier],
              col = .Options$DHARMaSignalColor, cex = 0.5)
 
@@ -269,7 +369,7 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
     else{
       a$col = checkDots("col", defaultCol, ...)
       a$pch = checkDots("pch", defaultPch, ...)
-      do.call(plot, append(list(res ~ pred, ylim = c(0,1), axes = FALSE), a))
+      do.call(plot, append(list(res ~ pred, ylim = c(0,1), axes = FALSE), a[names(a) != "main"]))
 
       axis(1)
       axis(2, at=c(0, quantiles, 1))
@@ -277,18 +377,17 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
 }
   ##### Quantile regressions #####
 
-  main = checkDots("main", ifelse(is.null(form), paste(yAxis, "vs. predicted"), paste(yAxis, "Residual vs. predictor")), ...)
+  if(is.null(a$main)) {main = ifelse(is.null(form), paste(yAxis, "vs. predicted"), paste(yAxis, "Residual vs. predictor"))} else {main = a$main}
   out = NULL
 
   if(is.numeric(pred)){
     if(quantreg == FALSE){
-      title(main = main, cex.main = 1)
-      abline(h = quantiles, col = "black", lwd = 0.5, lty = 2)
+      title(main = main, cex.main = 0.8)
       try({
         lines(smooth.spline(pred, res, df = 10), lty = 2, lwd = 2,
-              col = .Options$DHARMaSignalColor)
-        abline(h = 0.5, col = .Options$DHARMaSignalColor, lwd = 2)
+              col = "black")
       }, silent = TRUE)
+
     }else{
 
       out = testQuantiles(res, pred, quantiles = quantiles, plot = FALSE,
@@ -313,7 +412,7 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
       }
 
       title(main = main, cex.main = 0.8,
-            col.main = maincol)
+            col.main = maincol, line = 1)
 
       for(i in 1:length(quantiles)){
 
@@ -330,33 +429,6 @@ plotResiduals <- function(simulationOutput, form = NULL, quantreg = NULL,
     }
   }
   invisible(out)
-}
-
-
-
-
-#' Ensures the existence of a valid predictor to plot residuals against
-#'
-#' @param simulationOutput a DHARMa simulation output or an object that can be converted into a DHARMa simulation output.
-#' @param predictor an optional predictor. If no predictor is provided, will try to extract the fitted value.
-#' @keywords internal
-ensurePredictor <- function(simulationOutput,
-                            predictor = NULL){
-  if(!is.null(predictor)){
-
-    if(length(predictor) != length(simulationOutput$scaledResiduals)) stop("DHARMa: residuals and predictor do not have the same length. The issue is possibly that you have NAs in your predictor that were removed during the model fit. Remove the NA values from your predictor.")
-
-    if(is.character(predictor)) {
-      predictor = factor(predictor)
-      warning("DHARMa:::ensurePredictor: character string was provided as predictor. DHARMa has converted to factor automatically. To remove this warning, please convert to factor before attempting to plot with DHARMa.")
-    }
-
-  } else {
-
-    predictor = simulationOutput$fittedPredictedResponse
-    if(is.null(predictor)) stop("DHARMa: can't extract predictor from simulationOutput, and no predictor provided.")
-  }
-  return(predictor)
 }
 
 
@@ -377,3 +449,62 @@ plotConventionalResiduals <- function(fittedModel){
   plot(predict(fittedModel), resid(fittedModel, type = "response") , main = "Raw residuals" , ylab = "Residual", xlab = "Predicted")
   mtext("Conventional residual plots", outer = TRUE)
 }
+
+
+
+#' Plot residuals against all predictors of the model.
+#'
+#' Convenience function to plot residuals against all predictors of the model.
+#'
+#' @param simulationOutput an object with simulated residuals created by [simulateResiduals].
+#' @param predictorNames names of the predictors to be plotted.
+#' @param form optional predictor (and grouping variable) against which the residuals should be plotted.
+#' @param quantreg whether to perform a quantile regression or a smooth spline around the mean.
+#' @param rank if T, the values provided in form will be rank transformed.
+#' @param asFactor should a numeric predictor provided in form be treated as a factor.
+#' @param smoothScatter if T, a smooth scatter plot will plotted instead of a normal scatter plot.
+#' @param quantiles for a quantile regression, which quantiles should be plotted.
+#' @param absoluteDeviation if T, switch from displaying normal quantile residuals to absolute deviation from the mean expectation of 0.5 (calculated as 2 * abs(res - 0.5)).
+#' @param ... additional arguments to plot / boxplot.
+#' @details For further details on the parameters see [plotResiduals].
+#' @export
+#'
+plotResidualsAll <- function(simulationOutput, predictorNames, form = NULL, quantreg = NULL,
+                             rank = TRUE, asFactor = NULL, smoothScatter = NULL,
+                             quantiles = c(0.25, 0.5, 0.75),
+                             absoluteDeviation = FALSE, ...){
+
+  opar = par(no.readonly = TRUE)
+  on.exit(par(opar))
+
+  # get number of panels needed (copied from BayesianTools)
+  numPredictors = length(predictorNames)
+  if (numPredictors <= 0) stop("Unable to extract predictor names from model.")
+  lower = floor(sqrt(numPredictors))
+  upper = ceiling(sqrt(numPredictors))
+
+  if (lower == upper) {
+    mfrow = c(lower, lower)
+  } else if (lower * upper >= numPredictors) {
+    mfrow = c(lower, upper)
+  } else {
+    mfrow = c(upper, upper)
+  }
+
+  par(mfrow = mfrow)
+
+  # create residual plot for each predictor in the model (additional arguments in ... are ignored here)
+  for(i in 1:length(predictorNames)){
+    plotResiduals(simulationOutput,
+                  form = eval(formula(paste("~", predictorNames[i]))),
+                  quantreg = quantreg,
+                  rank = rank,
+                  asFactor = asFactor,
+                  smoothScatter = smoothScatter,
+                  quantiles = quantiles,
+                  absoluteDeviation = absoluteDeviation,
+                  xlab = predictorNames[i])
+  }
+
+}
+