+++ DHARMa/inst/examples/testQuantilesHelp.R
@@ -9,7 +9,7 @@ x # the test shows a combined p-value, corrected for multiple testing
 \dontrun{
 # accessing results of the test
 x$pvals # pvalues for the individual quantiles
-x$qgamFits # access the fitted quantile regression 
+x$qgamFits # access the fitted quantile regression
 summary(x$qgamFits[[1]]) # summary of the first fitted quantile
 
 # possible to test user-defined quantiles
@@ -18,6 +18,11 @@ testQuantiles(simulationOutput, quantiles = c(0.7))
 #  example with missing environmental predictor
 fittedModel <- glm(observedResponse ~ 1 , family = "poisson", data = testData)
 simulationOutput <- simulateResiduals(fittedModel = fittedModel)
+
+# specified as formula (recommended)
+testQuantiles(simulationOutput, predictor = ~Environment1)
+
+# or as variable in your environment
 testQuantiles(simulationOutput, predictor = testData$Environment1)
 
 plot(simulationOutput)